refactorings; phase #1
