+++ lib/nodejs_command_thread.py
